Grid development

+++ robotinterface/drivers/grbl/controller.py
@@ -54,7 +54,7 @@ class GrblDriver:
         logging.debug("Robot is idle again")
         return
 
-    async def move(self, x: int, y: int, z:int, feedrate: int) -> None:
+    async def move(self, x: float, y: float, z:float, feedrate: int) -> None:
         answer = await self._connection.send(f"G01 X{x} Y{y} Z{z} F{feedrate}")
         parser.move_parser(answer)
         await self._wait_till_idle()
